Modified index.js files
